refactor(frontend): move person-case session types
